requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.3.25
langchain-groq==0.3.2
langchain-core==0.3.59
python-dotenv==1.0.1
pydantic==2.7.4
